Create Default Subset Operators for Cohort Generation

This function configures a set of default subset operators used in cohort generation. It uses CohortGenerator package subset operator, and allows for easily specifying demographic and temporal parameters to subset a cohort and create a new cohort.

defaultCohortGeneratorSubsetOperator(
  minAge = 0,
  maxAge = 90,
  genderSubset = c(8532, 8507),
  restrictToEarliestOccurrence = TRUE,
  race = NULL,
  calendarStartDate = "2016-01-01",
  calendarEndDate = Sys.Date() - 365,
  priorTime = 365,
  followUpTime = 365,
  ethnicity = NULL,
  limitTo = "firstEver"
)

Arguments

minAge

Integer, optional. The minimum age at first drug exposure. Defaults to 0 years.

maxAge

Integer, optional. The maximum age at first drug exposure. Defaults to 90 years.

genderSubset

Numeric vector, optional. Identifies genders using concept IDs. Defaults to c(8532, 8507).

restrictToEarliestOccurrence

Logical, optional. Specifies whether to restrict the analysis to the earliest occurrence of the drug exposure. Defaults to TRUE.

race

Integer, optional. Concept ID for specifying race.

calendarStartDate

Date, optional. Specifies the earliest date from which to consider drug exposures, used to left-censor data. Defaults to '2016-01-01'.

calendarEndDate

Date, optional. The latest date up to which to consider drug exposures, used to right-censor data. By default, it is set to one year before the current date to account for claims accrual lag.

priorTime

Integer, optional. Represents the number of days of prior observation required before the start of drug exposure. Defaults to 365 days.

followUpTime

Integer, optional. Represents the number of days of follow-up observation after the start of drug exposure. Defaults to 365 days.

ethnicity

Integer, optional. Concept ID for specifying ethnicity.

limitTo

Character, optional. Describes whether to restrict the analysis to the first-ever observation period or to all observation periods with drug exposure. Defaults to 'firstEver'.

Value

A list of configured subset operators suitable for use in cohort generation functions.

Examples

defaultCohortGeneratorSubsetOperator(
  minAge = 18,
  maxAge = 65,
  genderSubset = c(8532),
  restrictToEarliestOccurrence = FALSE,
  race = 2106,
  calendarStartDate = as.Date("2010-01-01"),
  calendarEndDate = as.Date("2020-12-31"),
  priorTime = 180,
  followUpTime = 180,
  ethnicity = 38003564,
  limitTo = "all"
)
#> $limitValidGenderAgeMin0
#> <DemographicSubsetOperator>
#>   Inherits from: <SubsetOperator>
#>   Public:
#>     ageMax: active binding
#>     ageMin: active binding
#>     classname: function () 
#>     clone: function (deep = FALSE) 
#>     ethnicity: active binding
#>     gender: active binding
#>     getAutoGeneratedName: function () 
#>     getEthnicity: function () 
#>     getGender: function () 
#>     getQueryBuilder: function (id) 
#>     getRace: function () 
#>     initialize: function (definition = NULL) 
#>     isEqualTo: function (criteria) 
#>     mapGenderConceptsToNames: function (mapping = list(`8507` = "males", `8532` = "females", 
#>     name: active binding
#>     publicFields: function () 
#>     race: active binding
#>     toJSON: function () 
#>     toList: function () 
#>   Private:
#>     .ageMax: 65
#>     .ageMin: 18
#>     .ethnicity: 38003564
#>     .gender: 8532 8507
#>     .name: 
#>     .race: 2106
#>     baseFields: name
#>     queryBuilder: R6ClassGenerator
#>     suffixStr: Demo
#> 
#> $limitSubset
#> <LimitSubsetOperator>
#>   Inherits from: <SubsetOperator>
#>   Public:
#>     calendarEndDate: active binding
#>     calendarStartDate: active binding
#>     classname: function () 
#>     clone: function (deep = FALSE) 
#>     followUpTime: active binding
#>     getAutoGeneratedName: function () 
#>     getQueryBuilder: function (id) 
#>     initialize: function (definition = NULL) 
#>     isEqualTo: function (subsetOperatorB) 
#>     limitTo: active binding
#>     name: active binding
#>     priorTime: active binding
#>     publicFields: function () 
#>     toJSON: function () 
#>     toList: function () 
#>   Private:
#>     .calendarEndDate: NULL
#>     .calendarStartDate: NULL
#>     .followUpTime: 180
#>     .limitTo: all
#>     .name: 
#>     .priorTime: 180
#>     baseFields: name
#>     queryBuilder: R6ClassGenerator
#>     suffixStr: Limit to
#>